backend/requirements.txt
spotipy
python-dotenv
pymongo
